# openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: RESTful API for e-commerce application
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server

paths:
  /products:
    get:
      summary: Get all products
      description: Retrieve products with filtering, sorting, and pagination
      tags:
        - Products
      parameters:
        - name: search
          in: query
          description: Search in name, description, or brand
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          required: false
          schema:
            type: string
            enum: [Electronics, Clothing, Books, Home, Sports, Toys]
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            minimum: 0
        - name: sortBy
          in: query
          description: Sort option
          required: false
          schema:
            type: string
            enum: [newest, price-low, price-high, rating]
            default: newest
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products/{id}:
    get:
      summary: Get product by ID
      description: Retrieve a single product by its ID
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /products/categories/list:
    get:
      summary: Get all categories
      description: Retrieve list of all available product categories
      tags:
        - Products
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  enum: [Electronics, Clothing, Books, Home, Sports, Toys]
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders:
    post:
      summary: Create new order
      description: Create a new order (mock checkout)
      tags:
        - Orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /orders/{id}:
    get:
      summary: Get order by ID
      description: Retrieve order details by ID
      tags:
        - Orders
      parameters:
        - name: id
          in: path
          required: true
          description: Order ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /health:
    get:
      summary: Health check
      description: Check API health status
      tags:
        - Health
      responses:
        '200':
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: OK
                  message:
                    type: string
                    example: Server is running

components:
  schemas:
    Product:
      type: object
      required:
        - _id
        - name
        - description
        - price
        - image
        - category
        - inStock
        - stockCount
        - rating
        - numReviews
      properties:
        _id:
          type: string
          example: '507f1f77bcf86cd799439011'
        name:
          type: string
          example: 'Wireless Headphones'
        description:
          type: string
          example: 'High-quality wireless headphones with noise cancellation'
        price:
          type: number
          format: float
          minimum: 0
          example: 199.99
        image:
          type: string
          format: uri
          example: 'https://images.unsplash.com/photo-1505740420928-5e560c06d30e?w=500'
        category:
          type: string
          enum: [Electronics, Clothing, Books, Home, Sports, Toys]
          example: Electronics
        brand:
          type: string
          example: 'AudioTech'
        inStock:
          type: boolean
          example: true
        stockCount:
          type: integer
          minimum: 0
          example: 50
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.5
        numReviews:
          type: integer
          minimum: 0
          example: 120
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ProductsResponse:
      type: object
      required:
        - products
        - page
        - pages
        - total
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
          example: 1
        pages:
          type: integer
          example: 3
        total:
          type: integer
          example: 30

    OrderItem:
      type: object
      required:
        - product
        - name
        - image
        - price
        - quantity
      properties:
        product:
          type: string
          description: Product ID
          example: '507f1f77bcf86cd799439011'
        name:
          type: string
          example: 'Wireless Headphones'
        image:
          type: string
          format: uri
          example: 'https://images.unsplash.com/photo-1505740420928-5e560c06d30e?w=500'
        price:
          type: number
          format: float
          example: 199.99
        quantity:
          type: integer
          minimum: 1
          example: 2

    ShippingAddress:
      type: object
      required:
        - fullName
        - address
        - city
        - postalCode
        - country
      properties:
        fullName:
          type: string
          example: 'John Doe'
        address:
          type: string
          example: '123 Main St'
        city:
          type: string
          example: 'New York'
        postalCode:
          type: string
          example: '10001'
        country:
          type: string
          example: 'USA'

    CreateOrderRequest:
      type: object
      required:
        - orderItems
        - shippingAddress
        - paymentMethod
      properties:
        orderItems:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          enum: [Credit Card, PayPal, Cash on Delivery]
          example: Credit Card

    Order:
      type: object
      required:
        - _id
        - orderItems
        - shippingAddress
        - paymentMethod
        - itemsPrice
        - shippingPrice
        - taxPrice
        - totalPrice
        - isPaid
        - isDelivered
      properties:
        _id:
          type: string
          example: '507f1f77bcf86cd799439012'
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/ShippingAddress'
        paymentMethod:
          type: string
          enum: [Credit Card, PayPal, Cash on Delivery]
          example: Credit Card
        itemsPrice:
          type: number
          format: float
          example: 399.98
        shippingPrice:
          type: number
          format: float
          example: 10.0
        taxPrice:
          type: number
          format: float
          example: 40.0
        totalPrice:
          type: number
          format: float
          example: 449.98
        isPaid:
          type: boolean
          example: true
        isDelivered:
          type: boolean
          example: false
        paidAt:
          type: string
          format: date-time
        deliveredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Error message'
        error:
          type: object
          description: Detailed error (development only)